+++ CHANGELOG.md
@@ -66,9 +66,9 @@
 - [11577](https://github.com/vegaprotocol/vega/issues/11577) - Add API for party discounts and rewards.
 - [10716](https://github.com/vegaprotocol/vega/issues/10716) - Set Tendermint defaults during init.
 - [11612](https://github.com/vegaprotocol/vega/issues/11612) - Reward scaling support.
--[11624](https://github.com/vegaprotocol/vega/issues/11624) - prevent creation of rewards with no payout, but with high computational cost.
 - [11624](https://github.com/vegaprotocol/vega/issues/11624) - prevent creation of rewards with no payout, but with high computational cost.
 - [11512](https://github.com/vegaprotocol/vega/issues/11512) - Add loss socialisation amounts to funding payment API.
+- [11627](https://github.com/vegaprotocol/vega/issues/11627) - Add eligible keys filter to reward transfers. 
 
 ### 🐛 Fixes